+++ src/main/scala/io/viash/lenses/FunctionalityLenses.scala
@@ -26,6 +26,8 @@ object FunctionalityLenses {
   val nameLens = GenLens[Functionality](_.name)
   val versionLens = GenLens[Functionality](_.version)
   val requirementsLens = GenLens[Functionality](_.requirements)
+  val resourcesLens = GenLens[Functionality](_.resources)
+  val testResourcesLens = GenLens[Functionality](_.test_resources)
   val dependenciesLens = GenLens[Functionality](_.dependencies)
   val repositoriesLens = GenLens[Functionality](_.repositories)
 }