+++ src/test/scala/io/viash/functionality/dependencies/Dependency.scala
@@ -56,13 +56,123 @@ class DependencyTest extends AnyFunSuite with BeforeAndAfterAll {
     writeTestConfig(testFolder.resolve("src/dep2/config.vsh.yaml"), fun2)
 
     // build
-    val (stdout, stderr, exitCode) = TestHelper.testMainWithStdErr(
+    val testOutput = TestHelper.testMain(
         "ns", "build",
         "-s", testFolder.resolve("src").toString(),
         "-t", testFolder.resolve("target").toString()
       )
 
-    assert(stderr.strip == "All 2 configs built successfully", "check build was successful")
+    assert(testOutput.stderr.strip == "All 2 configs built successfully", "check build was successful")
+
+    // check file & file content
+    val outputPath = testFolder.resolve("target/executable/dep2/dep2")
+    val executable = outputPath.toFile
+    assert(executable.exists)
+    assert(executable.canExecute)
+
+    val outputText = IO.read(outputPath.toUri())
+    assert(outputText.contains("VIASH_DEP_DEP1="), "check the dependency is set in the output script")
+
+    // check output when running
+    val out = Exec.runCatch(
+      Seq(executable.toString)
+    )
+
+    assert(out.output == "Hello from dep1\nHello from dep2\n")
+    assert(out.exitValue == 0)
+  }
+
+  test("Use a local repository with an absolute path") {
+    val testFolder = createViashSubFolder(temporaryFolder, "local_test_absolute_path")
+    
+    // write test files
+    val fun1 = Functionality(
+      name = "dep1",
+      resources = textBashScript("echo Hello from dep1"),
+    )
+    val fun2 = Functionality(
+      name = "dep2",
+      resources = textBashScript("$dep_dep1\necho Hello from dep2"),
+      dependencies = List(Dependency("dep1", repository = Right(LocalRepository(path = Some("/dependencies")))))
+    )
+
+    writeTestConfig(testFolder.resolve("dependencies/src/dep1/config.vsh.yaml"), fun1)
+    writeTestConfig(testFolder.resolve("src/dep2/config.vsh.yaml"), fun2)
+
+    // build our local repository
+    val build1 = TestHelper.testMain(
+        workingDir = Some(testFolder.resolve("dependencies")),
+        "ns", "build",
+        "-s", testFolder.resolve("dependencies/src").toString(),
+        "-t", testFolder.resolve("dependencies/target").toString()
+      )
+    
+    assert(build1.stderr.strip == "All 1 configs built successfully", "check dependency build was successful")
+
+    // build
+    val build2 = TestHelper.testMain(
+        workingDir = Some(testFolder),
+        "ns", "build",
+        "-s", testFolder.resolve("src").toString(),
+        "-t", testFolder.resolve("target").toString()
+      )
+
+    assert(build2.stderr.strip == "All 1 configs built successfully", "check build was successful")
+
+    // check file & file content
+    val outputPath = testFolder.resolve("target/executable/dep2/dep2")
+    val executable = outputPath.toFile
+    assert(executable.exists)
+    assert(executable.canExecute)
+
+    val outputText = IO.read(outputPath.toUri())
+    assert(outputText.contains("VIASH_DEP_DEP1="), "check the dependency is set in the output script")
+
+    // check output when running
+    val out = Exec.runCatch(
+      Seq(executable.toString)
+    )
+
+    assert(out.output == "Hello from dep1\nHello from dep2\n")
+    assert(out.exitValue == 0)
+  }
+
+  test("Use a local repository with a relative path") {
+    val testFolder = createViashSubFolder(temporaryFolder, "local_test_absolute_path")
+    
+    // write test files
+    val fun1 = Functionality(
+      name = "dep1",
+      resources = textBashScript("echo Hello from dep1"),
+    )
+    val fun2 = Functionality(
+      name = "dep2",
+      resources = textBashScript("$dep_dep1\necho Hello from dep2"),
+      dependencies = List(Dependency("dep1", repository = Right(LocalRepository(path = Some("../../dependencies")))))
+    )
+
+    writeTestConfig(testFolder.resolve("dependencies/src/dep1/config.vsh.yaml"), fun1)
+    writeTestConfig(testFolder.resolve("src/dep2/config.vsh.yaml"), fun2)
+
+    // build our local repository
+    val build1 = TestHelper.testMain(
+        workingDir = Some(testFolder.resolve("dependencies")),
+        "ns", "build",
+        "-s", testFolder.resolve("dependencies/src").toString(),
+        "-t", testFolder.resolve("dependencies/target").toString()
+      )
+    
+    assert(build1.stderr.strip == "All 1 configs built successfully", "check dependency build was successful")
+
+    // build
+    val build2 = TestHelper.testMain(
+        workingDir = Some(testFolder),
+        "ns", "build",
+        "-s", testFolder.resolve("src").toString(),
+        "-t", testFolder.resolve("target").toString()
+      )
+
+    assert(build2.stderr.strip == "All 1 configs built successfully", "check build was successful")
 
     // check file & file content
     val outputPath = testFolder.resolve("target/executable/dep2/dep2")
@@ -95,13 +205,13 @@ class DependencyTest extends AnyFunSuite with BeforeAndAfterAll {
     writeTestConfig(testFolder.resolve("src/dep3/config.vsh.yaml"), fun)
 
     // build
-    val (stdout, stderr, exitCode) = TestHelper.testMainWithStdErr(
+    val testOutput = TestHelper.testMain(
         "ns", "build",
         "-s", testFolder.resolve("src").toString(),
         "-t", testFolder.resolve("target").toString()
       )
 
-    assert(stderr.strip.contains("All 1 configs built successfully"), "check build was successful")
+    assert(testOutput.stderr.strip.contains("All 1 configs built successfully"), "check build was successful")
 
     // check file & file content
     val outputPath = testFolder.resolve("target/executable/dep3/dep3")
@@ -134,13 +244,13 @@ class DependencyTest extends AnyFunSuite with BeforeAndAfterAll {
     writeTestConfig(testFolder.resolve("src/dep4/config.vsh.yaml"), fun)
 
     // build
-    val (stdout, stderr, exitCode) = TestHelper.testMainWithStdErr(
+    val testOutput = TestHelper.testMain(
         "ns", "build",
         "-s", testFolder.resolve("src").toString(),
         "-t", testFolder.resolve("target").toString()
       )
 
-    assert(stderr.strip.contains("All 1 configs built successfully"), "check build was successful")
+    assert(testOutput.stderr.strip.contains("All 1 configs built successfully"), "check build was successful")
 
     // check file & file content
     val outputPath = testFolder.resolve("target/executable/dep4/dep4")